--- src/io.cpp
@@ -1,74 +1,0 @@
-/* The MIT License
-
-   Copyright (c) 2016-2019 Karel Brinda (kbrinda@hsph.harvard.edu)
-
-   Permission is hereby granted, free of charge, to any person obtaining a copy
-   of this software and associated documentation files (the "Software"), to deal
-   in the Software without restriction, including without limitation the rights
-   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-   copies of the Software, and to permit persons to whom the Software is
-   furnished to do so, subject to the following conditions:
-
-   The above copyright notice and this permission notice shall be included in
-   all copies or substantial portions of the Software.
-
-   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-   SOFTWARE.
-*/
-
-#include "io.h"
-
-#include <cmath>
-#include <cstdarg>
-
-namespace ococo {
-
-void fatal_error(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:fatal-error]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void error(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:error]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void warning(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:warning]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void info(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-bool file_exists(const std::string &fn) {
-    FILE *file;
-
-    file = fopen(fn.c_str(), "r");
-    if (file) {
-        fclose(file);
-        return true;
-    }
-    return false;
-}
-
-}  // namespace ococo
